// src/webparts/helloWorld/components/HelloWorld.module.scss.d.ts
// Declare the ambient module for CSS classes
declare const styles: {
  helloWorld: string;
  teams: string;
  welcome: string;
  welcomeImage: string;
  links: string;
  customerForm: string;
  formRow: string;
  input: string;
  textarea: string;
  submitBtn: string;
  required: string;
  formError: string;
  fieldError: string;
};

export default styles;